// frontend/src/app/model/prescriptions/prescription.ts
export interface Prescription {
    populationGroup: string;
    chronicDisease: string;
    groupOfChronicDiseases: string;
    disease: string;
    level: string;
    exercise: string;
    modality: string;
    frequency: string;
    intensity: string;
    time: string;
    type: string;
    volume: string;
    progression: string;
    specialConsiderations: string;
}